Getting closer to at least a halfway working theme…
We are getting closer to having just a working, bare minimum theme.
Once we have this, we will work to get a live developer version of the
site up. Until then, if you want to keep track of progress, please
clone locally, install wordpress using a tool like WAMP/MAMP, and
install this theme.